requirements.txt
accelerate==1.13.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
av==17.0.0
cachetools
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
defusedxml==0.7.1
filelock==3.25.2
fonttools==4.62.1
fpdf2==2.8.7
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.10.2
idna==3.11
jinja2==3.1.6
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.3
ollama==0.6.1
reportlab==4.4.10
onnxruntime-gpu
opencv-python
packaging==26.1
pdf2image==1.17.0
pdfplumber
pillow==12.1.1
psutil==7.2.2
pyclipper
pydantic==2.13.1
pydantic-core==2.46.1
pygments==2.20.0
pyyaml==6.0.3
qwen-vl-utils==0.0.14
regex==2026.4.4
requests==2.33.1
rich==15.0.0
ruamel.yaml
safetensors==0.7.0
setuptools==70.2.0
shapely
shellingham==1.5.4
six
sympy==1.14.0
tokenizers==0.22.2
torch==2.11.0+cu128
torchaudio==2.11.0+cu128
torchvision==0.26.0+cu128
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
vietocr==0.3.13
